Fix evaluating within multiple environments
When running a program that pushes to the `EnvironmentStack` mulitple
times (with recursion) we always retrive values from the first one since
we were pushing them to the back rather than the front, which would
always fetch the original value.

This changes the `Environment` to use a `VecDeque` instead and always
pushes new environments to the front rather than the back.